Откорректировать код сервисов, чтобы репозитории не ссылались друг на друга, а взаимодействвовали в сервисе.
